+++ saleor/plugins/avatax/plugin.py
@@ -71,6 +71,8 @@ class AvataxPlugin(BasePlugin):
         {"name": "from_country_area", "value": None},
         {"name": "from_postal_code", "value": None},
         {"name": "shipping_tax_code", "value": "FR000000"},
+        {"name": "override_global_tax", "value": False},
+        {"name": "include_taxes_in_prices", "value": True},
     ]
     CONFIG_STRUCTURE = {
         "Username or account": {
@@ -136,6 +138,16 @@ class AvataxPlugin(BasePlugin):
             ),
             "label": "Shipping tax code",
         },
+        "override_global_tax": {
+            "type": ConfigurationTypeField.BOOLEAN,
+            "help_text": "Used when setting per channel is needed.",
+            "label": "Override global tax settings.",
+        },
+        "include_taxes_in_prices": {
+            "type": ConfigurationTypeField.BOOLEAN,
+            "help_text": 'Applied only if "Override global tax settings" is on.',
+            "label": "All products prices are entered with tax included.",
+        },
     }
 
     def __init__(self, *args, **kwargs):
@@ -158,6 +170,8 @@ class AvataxPlugin(BasePlugin):
             from_country_area=configuration["from_country_area"],
             from_postal_code=configuration["from_postal_code"],
             shipping_tax_code=configuration["shipping_tax_code"],
+            override_global_tax=configuration["override_global_tax"],
+            include_taxes_in_prices=configuration["include_taxes_in_prices"],
         )
 
     def _skip_plugin(
@@ -194,13 +208,11 @@ class AvataxPlugin(BasePlugin):
 
         if not _validate_checkout(checkout_info, lines):
             return checkout_total
-        response = get_checkout_tax_data(checkout_info, lines, discounts, self.config)
+        response = get_checkout_tax_data(
+            checkout_info, lines, self.config.tax_included, discounts, self.config
+        )
         if not response or "error" in response:
             return checkout_total
-
-        tax_included = (
-            lambda: Site.objects.get_current().settings.include_taxes_in_prices
-        )
 
         currency = checkout_info.checkout.currency
         taxed_total = zero_taxed_money(currency)
@@ -209,7 +221,7 @@ class AvataxPlugin(BasePlugin):
             taxed_line_total_data = self._calculate_checkout_line_total_price(
                 taxes_data=response,
                 item_code=line.variant.sku or line.variant.get_global_id(),
-                tax_included=tax_included,
+                tax_included=self.config.tax_included,
                 # for some cases we will need a base_value but no need to call it for
                 # each line
                 base_value=SimpleLazyObject(  # type:ignore
@@ -250,10 +262,7 @@ class AvataxPlugin(BasePlugin):
                 shipping_tax = Decimal(line["tax"])
                 break
 
-        tax_included = (
-            lambda: Site.objects.get_current().settings.include_taxes_in_prices
-        )
-        if currency == "JPY" and tax_included():
+        if currency == "JPY" and self.config.tax_included:
             shipping_gross = Money(amount=shipping_price.amount, currency=currency)
             shipping_net = Money(
                 amount=shipping_gross.amount - shipping_tax, currency=currency
@@ -285,7 +294,9 @@ class AvataxPlugin(BasePlugin):
         if not _validate_checkout(checkout_info, lines):
             return base_shipping_price
 
-        response = get_checkout_tax_data(checkout_info, lines, discounts, self.config)
+        response = get_checkout_tax_data(
+            checkout_info, lines, self.config.tax_included, discounts, self.config
+        )
         if not response or "error" in response:
             return base_shipping_price
 
@@ -320,6 +331,7 @@ class AvataxPlugin(BasePlugin):
             checkout_info,
             lines,
             self.config,
+            self.config.tax_included,
             transaction_token=str(checkout_info.checkout.token),
             transaction_type=TransactionType.ORDER,
             discounts=discounts,
@@ -353,7 +365,9 @@ class AvataxPlugin(BasePlugin):
     def order_created(self, order: "Order", previous_value: Any) -> Any:
         if not self.active or order.is_unconfirmed():
             return previous_value
-        request_data = get_order_request_data(order, self.config)
+        request_data = get_order_request_data(
+            order, self.config, self.config.tax_included
+        )
 
         transaction_url = urljoin(
             get_api_url(self.config.use_sandbox), "transactions/createoradjust"
@@ -385,17 +399,15 @@ class AvataxPlugin(BasePlugin):
         if not _validate_checkout(checkout_info, lines):
             return base_total
 
-        tax_included = (
-            lambda: Site.objects.get_current().settings.include_taxes_in_prices
+        taxes_data = get_checkout_tax_data(
+            checkout_info, lines, self.config.tax_included, discounts, self.config
         )
-
-        taxes_data = get_checkout_tax_data(checkout_info, lines, discounts, self.config)
         variant = checkout_line_info.variant
 
         return self._calculate_checkout_line_total_price(
             taxes_data,
             variant.sku or variant.get_global_id(),
-            tax_included,
+            self.config.tax_included,
             previous_value,
         )
 
@@ -403,7 +415,7 @@ class AvataxPlugin(BasePlugin):
     def _calculate_checkout_line_total_price(
         taxes_data: Dict[str, Any],
         item_code: str,
-        tax_included: Callable[[], bool],
+        tax_included: bool,
         base_value: TaxedMoney,
     ) -> TaxedMoney:
         if not taxes_data or "error" in taxes_data:
@@ -421,7 +433,7 @@ class AvataxPlugin(BasePlugin):
             discount_amount = Decimal(line.get("discountAmount", 0.0))
             net = Decimal(line["lineAmount"])
 
-            if currency == "JPY" and tax_included():
+            if currency == "JPY" and tax_included:
                 line_gross = base_value.gross
                 line_net = Money(amount=line_gross.amount - tax, currency=currency)
             else:
@@ -449,15 +461,11 @@ class AvataxPlugin(BasePlugin):
         if not _validate_order(order):
             return previous_value
 
-        tax_included = (
-            lambda: Site.objects.get_current().settings.include_taxes_in_prices
-        )
-
         taxes_data = self._get_order_tax_data(order, previous_value)
         return self._calculate_order_line_total_price(
             taxes_data,
             variant.sku or variant.get_global_id(),
-            tax_included,
+            self.config.tax_included,
             previous_value,
         )
 
@@ -465,7 +473,7 @@ class AvataxPlugin(BasePlugin):
     def _calculate_order_line_total_price(
         taxes_data: Dict[str, Any],
         item_code: str,
-        tax_included: Callable[[], bool],
+        tax_included: bool,
         base_value: OrderTaxedPricesData,
     ) -> OrderTaxedPricesData:
         if not taxes_data or "error" in taxes_data:
@@ -484,7 +492,7 @@ class AvataxPlugin(BasePlugin):
             discount_amount = Decimal(line.get("discountAmount", 0.0))
             net = Decimal(line["lineAmount"])
 
-            if currency == "JPY" and tax_included():
+            if currency == "JPY" and tax_included:
                 line_gross = Money(
                     base_value.price_with_discounts.gross.amount - discount_amount,
                     currency,
@@ -524,18 +532,17 @@ class AvataxPlugin(BasePlugin):
         if not _validate_checkout(checkout_info, lines):
             return base_total
 
-        tax_included = (
-            lambda: Site.objects.get_current().settings.include_taxes_in_prices
-        )
         variant = checkout_line_info.variant
 
         quantity = checkout_line_info.line.quantity
-        taxes_data = get_checkout_tax_data(checkout_info, lines, discounts, self.config)
+        taxes_data = get_checkout_tax_data(
+            checkout_info, lines, self.config.tax_included, discounts, self.config
+        )
         default_total = previous_value * quantity
         taxed_total_price = self._calculate_checkout_line_total_price(
             taxes_data,
             variant.sku or variant.get_global_id(),
-            tax_included,
+            self.config.tax_included,
             default_total,
         )
         return taxed_total_price / quantity
@@ -551,10 +558,6 @@ class AvataxPlugin(BasePlugin):
         if not variant or (variant and not product.charge_taxes):
             return previous_value
 
-        tax_included = (
-            lambda: Site.objects.get_current().settings.include_taxes_in_prices
-        )
-
         quantity = order_line.quantity
         taxes_data = self._get_order_tax_data(order, previous_value)
         default_total = OrderTaxedPricesData(
@@ -564,7 +567,7 @@ class AvataxPlugin(BasePlugin):
         taxed_total_prices_data = self._calculate_order_line_total_price(
             taxes_data,
             variant.sku or variant.get_global_id(),
-            tax_included,
+            self.config.tax_included,
             default_total,
         )
         return OrderTaxedPricesData(
@@ -584,17 +587,16 @@ class AvataxPlugin(BasePlugin):
 
         if not _validate_order(order):
             return previous_value
-        taxes_data = get_order_tax_data(order, self.config, False)
-
-        tax_included = (
-            lambda: Site.objects.get_current().settings.include_taxes_in_prices
+        taxes_data = get_order_tax_data(
+            order, self.config, self.config.tax_included, False
         )
+
         currency = taxes_data.get("currencyCode")
         for line in taxes_data.get("lines", []):
             if line["itemCode"] == "Shipping":
                 tax = Decimal(line.get("tax", 0.0))
                 net = Decimal(line.get("lineAmount", 0.0))
-                if currency == "JPY" and tax_included():
+                if currency == "JPY" and self.config.tax_included:
                     gross = previous_value.gross
                     net = Money(amount=gross.amount - tax, currency=currency)
                 else:
@@ -692,7 +694,7 @@ class AvataxPlugin(BasePlugin):
             return None
 
         response = get_checkout_tax_data(
-            checkout_info, lines_info, discounts, self.config
+            checkout_info, lines_info, self.config.tax_included, discounts, self.config
         )
         if not response or "error" in response:
             return None
@@ -709,7 +711,9 @@ class AvataxPlugin(BasePlugin):
         if not valid:
             return None
 
-        response = get_order_tax_data(order, self.config, False)
+        response = get_order_tax_data(
+            order, self.config, self.config.tax_included, False
+        )
         if not response or "error" in response:
             return None